+++ OpenVBX/libraries/User_Controller.php
@@ -367,7 +367,7 @@ class User_Controller extends MY_Controller
 		try
 		{
 			/* Retrieve twilio numbers w/o sandbox */
-			$numbers = $this->vbx_incoming_numbers->get_numbers(false);
+			$numbers = $this->vbx_incoming_numbers->get_numbers();
 		}
 		catch(VBX_IncomingNumberException $e)
 		{